add css + about page

+++ app/models/user.rb
@@ -11,11 +11,11 @@ class User < ActiveRecord::Base
     "Anonymous_#{id}"
   end
 
-  def like_post(post_id)
-    post_likes.find_or_create_by(post_id: post_id)
-  end
-
-  def dislike_post(post_id)
-    post_likes.where(post_id: post_id).destroy_all
+  def like_post(post_id, liked)
+    if liked
+      post_likes.find_or_create_by(post_id: post_id)
+    else
+      post_likes.where(post_id: post_id).destroy_all
+    end
   end
 end
